--- EssentialFeed/EssentialFeedTests/Image Comments API/LoadImageCommentsFromRemoteUseCaseTests.swift
@@ -1,137 +1,0 @@
-//
-//  LoadImageCommentsFromRemoteUseCaseTests.swift
-//  EssentialFeedTests
-//
-//  Created by Prabhat Tiwari on 15/09/25.
-//
-
-import XCTest
-import EssentialFeed
-
-class LoadImageCommentsFromRemoteUseCaseTests: XCTestCase {
-    
-    func test_load_deliversErrorOnNon200HTTPResponse() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [199, 150, 300, 400, 500]
-            
-        samples.enumerated().forEach { index, code in
-            expect(sut, completeWith: failure(.invalidData), when: {
-                let data = makeItemJSON([])
-                client.complete(withStatusCode: code, data: data, at: index)
-            })
-        }
-        
-    }
-    
-    func test_load_deliversErrorOn2xxHTTPResponseWithInvalidJSON() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [200, 201, 250, 210, 299]
-        
-        samples.enumerated().forEach { index, code in
-            expect(sut, completeWith: failure(.invalidData), when: {
-                let invalidJson = Data("Invlaid JSON".utf8)
-                client.complete(withStatusCode: code, data: invalidJson, at: index)
-            })
-        }
-    }
-    
-    func test_load_deliverNoItemsOn2xxHTTPResponseWithEmptyJSONList() {
-        let (sut, client) = makeSUT()
-        
-        let samples = [200, 201, 250, 210, 299]
-        
-        samples.enumerated().forEach { index, code in
-            expect(sut, completeWith: .success([]), when: {
-                let data = makeItemJSON([])
-                client.complete(withStatusCode: code, data: data, at: index)
-            })
-        }
-    }
-    
-    func test_load_deliverItemsOn2xxHTTPResponseWithJSONList() {
-        let (sut, client) = makeSUT()
-        
-        let item1 = makeItem(
-            id: UUID(),
-            message: "a message",
-            createdAt: (Date(timeIntervalSince1970: 1598627222), "2020-08-28T15:07:02+00:00"),
-            username: "a username")
-        
-        let item2 = makeItem(
-            id: UUID(),
-            message: "another message",
-            createdAt: (Date(timeIntervalSince1970: 1577881882), "2020-01-01T12:31:22+00:00"),
-            username: "another username")
-        
-        let items = [item1.model, item2.model]
-        
-        
-        let samples = [200, 201, 250, 210, 299]
-        
-        samples.enumerated().forEach { index, code in
-            expect(sut, completeWith: .success(items), when: {
-                let data = makeItemJSON([item1.json, item2.json])
-                client.complete(withStatusCode: code, data: data, at: index)
-            })
-        }
-    
-    }
-    
-    //MARK: - Helpers
-    
-    private func makeSUT(url: URL = URL(string: "https://a-url.com")!, file: StaticString = #filePath, line: UInt = #line) -> (RemoteImageCommentsLoader, HTTPClientSpy) {
-        let client = HTTPClientSpy()
-        let sut = RemoteImageCommentsLoader(url: url, client: client)
-        trackForMemoryLeaks(client)
-        trackForMemoryLeaks(sut)
-        return (sut, client)
-    }
-    
-    private func failure(_ error: RemoteImageCommentsLoader.Error) -> RemoteImageCommentsLoader.Result {
-        .failure(error)
-    }
-    
-    private func makeItem(id: UUID, message: String, createdAt: (date: Date, iso8601String: String), username: String) -> (model: ImageComment, json: [String: Any]) {
-        let item = ImageComment(id: id, message: message, createdAt: createdAt.date, username: username)
-        
-        let json: [String: Any] = [
-            "id": id.uuidString,
-            "message": message,
-            "created_at": createdAt.iso8601String,
-            "author": [
-                "username": username
-            ]
-        ]
-        
-        return (item, json)
-    }
-    
-    private func makeItemJSON(_ items:[[String: Any]]) -> Data {
-        let json = ["items": items]
-        return try! JSONSerialization.data(withJSONObject: json)
-    }
-    
-    private func expect(_ sut: RemoteImageCommentsLoader, completeWith expectedResult: RemoteImageCommentsLoader.Result, when action: () -> Void, file: StaticString = #filePath, line: UInt = #line) {
-        let exp = expectation(description: "wait for load completion")
-        
-        sut.load { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedItems), .success(expectedItems)):
-                XCTAssertEqual(receivedItems, expectedItems, file: file, line: line)
-            case let (.failure(receivedError as RemoteImageCommentsLoader.Error), .failure(expectedError as RemoteImageCommentsLoader.Error)):
-                XCTAssertEqual(receivedError, expectedError, file: file, line: line)
-            default:
-              XCTFail("Expected result \(expectedResult) got \(receivedResult) instead", file: file, line: line)
-            }
-            exp.fulfill()
-        }
-        
-        action()
-        
-        wait(for: [exp], timeout: 1.0)
-        
-    }
-    
-}
